Speedup SEMFEM assembly
