+++ config/elliptic.cmake
@@ -4,7 +4,7 @@ set(ELLIPTIC_SOURCES
         ${ELLIPTIC_SOURCE_DIR}/linearSolver/PCG.cpp
         ${ELLIPTIC_SOURCE_DIR}/linearSolver/PGMRES.cpp
         ${ELLIPTIC_SOURCE_DIR}/amgSolver/amgx/amgx.c
-        ${ELLIPTIC_SOURCE_DIR}/fem_amg_preco.c
+        ${ELLIPTIC_SOURCE_DIR}/fem_amg_preco.cpp
         ${ELLIPTIC_SOURCE_DIR}/ellipticBuildContinuous.cpp
         ${ELLIPTIC_SOURCE_DIR}/ellipticBuildContinuousGalerkin.cpp
         ${ELLIPTIC_SOURCE_DIR}/ellipticJacobi.cpp
